+++ src/pybnb/common.py
@@ -124,6 +124,9 @@ class TerminationCondition(str, enum.Enum):
     queue_empty = "queue_empty"
     """The dispatcher terminated the solve due to the node
     queue becoming empty."""
+    queue_limit = "queue_limit"
+    """The dispatcher terminated the solve due to the
+    user-supplied queue size limit being exceeded."""
     interrupted = "interrupted"
     """Solve termination was initiated by SIGINT or SIGUSR
     signal event."""