+++ src/tests/mpi/test_misc.py
@@ -60,7 +60,7 @@ def _logging_check(comm):
                              initialize_queue,
                              "bound",
                              ConvergenceChecker(p.sense()),
-                             None, None,
+                             None, None, None,
                              get_simple_logger(console=True,
                                                stream=out,
                                                level=logging.DEBUG,