+++ src/tests/test_dispatcher.py
@@ -55,6 +55,7 @@ class TestDispatcherSimple(object):
     def test_queue_strategy(self):
         node_limit = None
         time_limit = None
+        queue_limit = None
         log = get_simple_logger()
         log_interval_seconds = inf
         log_new_incumbent = True
@@ -78,6 +79,7 @@ class TestDispatcherSimple(object):
             convergence_checker,
             node_limit,
             time_limit,
+            queue_limit,
             log,
             log_interval_seconds,
             log_new_incumbent)
@@ -90,6 +92,7 @@ class TestDispatcherSimple(object):
             convergence_checker,
             node_limit,
             time_limit,
+            queue_limit,
             log,
             log_interval_seconds,
             log_new_incumbent)
@@ -102,6 +105,7 @@ class TestDispatcherSimple(object):
             convergence_checker,
             node_limit,
             time_limit,
+            queue_limit,
             log,
             log_interval_seconds,
             log_new_incumbent)
@@ -114,6 +118,7 @@ class TestDispatcherSimple(object):
             convergence_checker,
             node_limit,
             time_limit,
+            queue_limit,
             log,
             log_interval_seconds,
             log_new_incumbent)
@@ -126,6 +131,7 @@ class TestDispatcherSimple(object):
             convergence_checker,
             node_limit,
             time_limit,
+            queue_limit,
             log,
             log_interval_seconds,
             log_new_incumbent)
@@ -138,6 +144,7 @@ class TestDispatcherSimple(object):
             convergence_checker,
             node_limit,
             time_limit,
+            queue_limit,
             log,
             log_interval_seconds,
             log_new_incumbent)
@@ -150,6 +157,7 @@ class TestDispatcherSimple(object):
             convergence_checker,
             node_limit,
             time_limit,
+            queue_limit,
             log,
             log_interval_seconds,
             log_new_incumbent)
@@ -162,6 +170,7 @@ class TestDispatcherSimple(object):
             convergence_checker,
             node_limit,
             time_limit,
+            queue_limit,
             log,
             log_interval_seconds,
             log_new_incumbent)
@@ -174,6 +183,7 @@ class TestDispatcherSimple(object):
             convergence_checker,
             node_limit,
             time_limit,
+            queue_limit,
             log,
             log_interval_seconds,
             log_new_incumbent)
@@ -186,6 +196,7 @@ class TestDispatcherSimple(object):
             convergence_checker,
             node_limit,
             time_limit,
+            queue_limit,
             log,
             log_interval_seconds,
             log_new_incumbent)
@@ -194,6 +205,7 @@ class TestDispatcherSimple(object):
     def test_initialize_queue(self):
         node_limit = None
         time_limit = None
+        queue_limit = None
         log = get_simple_logger()
         log_interval_seconds = inf
         log_new_incumbent = True
@@ -220,6 +232,7 @@ class TestDispatcherSimple(object):
             convergence_checker,
             node_limit,
             time_limit,
+            queue_limit,
             log,
             log_interval_seconds,
             log_new_incumbent)
@@ -234,6 +247,7 @@ class TestDispatcherSimple(object):
             convergence_checker,
             node_limit,
             time_limit,
+            queue_limit,
             log,
             log_interval_seconds,
             log_new_incumbent)
@@ -249,6 +263,7 @@ class TestDispatcherSimple(object):
             convergence_checker,
             node_limit,
             time_limit,
+            queue_limit,
             log,
             log_interval_seconds,
             log_new_incumbent)
@@ -265,6 +280,7 @@ class TestDispatcherSimple(object):
             convergence_checker,
             node_limit,
             time_limit,
+            queue_limit,
             log,
             log_interval_seconds,
             log_new_incumbent)
@@ -281,6 +297,7 @@ class TestDispatcherSimple(object):
             convergence_checker,
             node_limit,
             time_limit,
+            queue_limit,
             log,
             log_interval_seconds,
             log_new_incumbent)
@@ -301,6 +318,7 @@ class TestDispatcherSimple(object):
                 convergence_checker,
                 node_limit,
                 time_limit,
+                queue_limit,
                 log,
                 log_interval_seconds,
                 log_new_incumbent)