+++ CHANGELOG.rst
@@ -5,6 +5,9 @@ Changelog
 ~~~~~~~~~~~~~~~~~
 
 * Renaming "no_nodes" termination condition to "queue_empty"
+* Adding a `queue_limit` solve option, which causes the
+  solve to terminate with the "queue_limit" termination
+  condition when the queue size exceeds this value.
 
 0.5.2 - 2019-02-13
 ~~~~~~~~~~~~~~~~~~